Mudança Id dos Beans

+++ src/main/java/br/edu/utfpr/cm/pi/daos/CargoDao.java
@@ -2,7 +2,7 @@ package br.edu.utfpr.cm.pi.daos;
 
 import br.edu.utfpr.cm.pi.beans.Cargo;
 
-public class CargoDao extends JpaDao<Cargo, Integer> {
+public class CargoDao extends JpaDao<Cargo, Long> {
 
 	public CargoDao() {
 		super(Cargo.class);
