+++ src/main/java/br/edu/utfpr/cm/pi/daos/FuncionarioDao.java
@@ -2,7 +2,7 @@ package br.edu.utfpr.cm.pi.daos;
 
 import br.edu.utfpr.cm.pi.beans.Funcionario;
 
-public class FuncionarioDao extends JpaDao<Funcionario, Integer> {
+public class FuncionarioDao extends JpaDao<Funcionario, Long> {
 
 	public FuncionarioDao() {
 		super(Funcionario.class);
