+++ src/main/java/br/edu/utfpr/cm/pi/daos/TransacaoDao.java
@@ -2,7 +2,7 @@ package br.edu.utfpr.cm.pi.daos;
 
 import br.edu.utfpr.cm.pi.beans.Transacao;
 
-public class TransacaoDao extends JpaDao<Transacao, Integer> {
+public class TransacaoDao extends JpaDao<Transacao, Long> {
 
 	public TransacaoDao() {
 		super(Transacao.class);
