+++ src/main/java/br/edu/utfpr/cm/pi/daos/UsuarioDao.java
@@ -2,7 +2,7 @@ package br.edu.utfpr.cm.pi.daos;
 
 import br.edu.utfpr.cm.pi.beans.Usuario;
 
-public class UsuarioDao extends JpaDao<Usuario, Integer> {
+public class UsuarioDao extends JpaDao<Usuario, Long> {
 
 	public UsuarioDao() {
 		super(Usuario.class);
